renamed files to .cjs so they can be reused in the grcp esm tests
