+++ test/versioned/grpc/server-client-streaming.tap.js
@@ -9,7 +9,7 @@ const tap = require('tap')
 const helper = require('../../lib/agent_helper')
 const DESTINATIONS = require('../../../lib/config/attribute-filter').DESTINATIONS
 const DESTINATION = DESTINATIONS.TRANS_EVENT | DESTINATIONS.ERROR_EVENT
-const { ERR_CODE, ERR_SERVER_MSG, HALT_CODE, HALT_GRPC_SERVER_MSG } = require('./constants')
+const { ERR_CODE, ERR_SERVER_MSG, HALT_CODE, HALT_GRPC_SERVER_MSG } = require('./constants.cjs')
 
 const {
   makeClientStreamingRequest,
@@ -20,7 +20,7 @@ const {
   assertServerTransaction,
   assertServerMetrics,
   assertDistributedTracing
-} = require('./util')
+} = require('./util.cjs')
 
 tap.test('gRPC Server: Client Streaming', (t) => {
   t.autoend()
